refactor(phase-4): split twitchextv2 into extensions/twitch/ package
Break the 1553-line twitchextv2.py monolith into a package of four feature
mixins + a thin glue Extension class, following the target layout from
tracking issue #15.

Before
- extensions/twitchextv2.py ........................ 1553 lines

After
- extensions/twitch/__init__.py .................... 341  (glue: lifecycle +
                                                            slash command +
                                                            streamer registry)
- extensions/twitch/_common.py .....................  55  (StreamerInfo,
                                                            ensure_utc)
- extensions/twitch/notifications.py ...............  311  (embed builders +
                                                             start/end/update
                                                             notifications)
- extensions/twitch/eventsub.py ....................  247  (websocket run/
                                                             revocation +
                                                             on_live_start/end/
                                                             on_update handlers)
- extensions/twitch/schedule.py ....................  243  (schedule embed,
                                                             planning message,
                                                             scheduled event
                                                             sync, update Task)
- extensions/twitch/emotes.py ......................  348  (emote cache +
                                                             diff detection
                                                             Task)

The single TwitchExtension class remains — feature areas are mixed in via MRO
(NotificationsMixin, EventSubMixin, ScheduleMixin, EmotesMixin) so
@slash_command / @listen / @Task.create decorators stay with the class that
holds the logic. No behaviour change, no public API change.

Auto-discovery plumbing
- main.py: extend the extensions/ scan so a directory with __init__.py is
  loaded as a package (previously only *.py files were picked up).
- src/webui/app.py: same for the config-module ↔ extension mapping used by
  WebUI auto-reload after config saves — walk every .py inside a package so
  load_config("...") calls in submodules are still discovered.

Verification
- python -m compileall extensions/ main.py src/ (src/raiderio.py skipped,
  unrelated py3.12-only f-string syntax; same skip as earlier phases).
- ruff check + ruff format --check clean across the new package and the two
  edited files.

Refs #15 (Phase 4 — God-extension split, 1/6 complete).
